Update master.py
Found the bug in the auto shutdown. The pi does not have CMOS time so it has to get it from the NTP. I think the time that is passed in the morning when the bug is exhibited is the time from the night before when it last shut down. I have added a minimum 10 min up-time before the 10 min delay before shutdown that should stop this glitch.